Make endpoint optional
The endpoint parameter is only required if wireguard should connect to a
remote site. That's not always possible. For example in situations where
the remote site is behind a NAT gateway and/or has a dynamic IP address.
For such setups you can create a 'passive' configuration that listens
for incoming packets.