Make filtering on dest addr optional
Previously `$destination_addresses` was an array with the public
ipv6/ipv4 addresses. Those were used as destination address in the
firewall rules. People might have dynamic local ips so this won't
work/creates a lot of noise due to constant firewall rule updates. As a
fix, it was already possible to set `$destination_addresses` to an empty
Array (`[]`). But passing this to the `daddr` parameter in ferm::rule
created a broken firewall rule. This patch sets the `daddr` parameter to
`undef` if `$destination_addresses` is an empty array.