fact: support keyfiles with multiple dots
Previously we assumed that the path to a key always contains an exact
amount of slashes and the filename only a single dot. Although the path
is still hardcoded, we now select the filename from the path and trim
the last `.pub`. People can now use multiple dots in their key names.