Make src port for firwalling configureable
Usually we parse the src port for incoming packets / the dst port for outgoing packets from the $endpoint param The param is optional, in case you want to create a passive endpoint for clients with dynamic ip addresses In those cases we still need to create firewall rules, but without src port for incoming packets / the dst port To make this all a bit easier, we also added a new parameter, $endpoint_port, which takes precedence over parsing $endpoint.

Previously the catalog compilation failed with:
```
Evaluation Error: Left match operand must result in a String value. Got an Undef Value.
```

Because of: `if $endpoint =~  /:(\d+)$/ {`

https://github.com/voxpupuli/puppet-wireguard/pull/103 verifies that the
current code is broken.

Thew new code adds a safeguard to check if $endpoint is set. we also add
a new parameter in case we want to explicitly set the port.