Ensure that we create firewall rules when destination_addresses is empty
Previously we first iterated on source_addresses, within that on
destination_addresses. That won't create firewall rules when
source_addresses is empty. That's the case for passive servers. We need
to flip the logic because we always have destination_addresses. We first
need to iterate on them. That ensures we always have firewall rules,
with optional source ip filtering.